requirements.txt
pygame

Pillow
